+++ NEWS.md
@@ -4,6 +4,7 @@
 * Expand functionality of `gt_group()` to allow `gt_group` objects to be combined with `gt_tbls` (#2128)
 * gt no longer exports `%>%`. You may either switch to the base pipe or call `library(magrittr)`. (#2150)
 * gt no longer export deprecated `dplyr::vars()` and `dplyr::one_of()`. (#2150)
+* Ensure `keep_with_next` parameter is applied consistently in `as_word` (#2149)
 
 # gt 1.3.0